backend/requirements.txt
Django==2.0.5
django-cors-headers==2.2.0
django-ordered-model==1.4.3
djangorestframework==3.8.2
pytz==2018.4
